requirements.txt
python-telegram-bot[job-queue]>=20.0
openai>=1.0
feedparser>=6.0
httpx>=0.25
beautifulsoup4>=4.12
python-dotenv>=1.0
apscheduler>=3.10
